+++ lib/app_state.dart
@@ -1,16 +1,21 @@
-export 'src/blocs/page_stack/back_button_dispatcher.dart'   show PageStackBackButtonDispatcher;
-export 'src/blocs/page_stack/event.dart'                    show PageStackBlocEvent;
-export 'src/blocs/page_stack/page_stack.dart'               show PageStackBloc;
-export 'src/blocs/page_stack/screen_event.dart'             show PageStackScreenBlocEvent;
+export 'src/blocs/app/app.dart'                                     show AppBloc;
+export 'src/blocs/app/normalized_state.dart'                        show AppBlocNormalizedState;
 
-export 'src/blocs/screen/close_event.dart'                  show ScreenBlocCloseEvent;
-export 'src/blocs/screen/configuration_changed_event.dart'  show ScreenBlocConfigurationChangedEvent;
-export 'src/blocs/screen/event.dart'                        show ScreenBlocEvent;
-export 'src/blocs/screen/screen.dart'                       show ScreenBloc;
+export 'src/blocs/page_stack/back_button_dispatcher.dart'           show PageStackBackButtonDispatcher;
+export 'src/blocs/page_stack/event.dart'                            show PageStackBlocEvent;
+export 'src/blocs/page_stack/page_stack.dart'                       show PageStackBloc;
+export 'src/blocs/page_stack/page_stack_bloc_normalized_state.dart' show PageStackBlocNormalizedState;
+export 'src/blocs/page_stack/screen_bloc_normalized_state.dart'     show ScreenBlocNormalizedState;
+export 'src/blocs/page_stack/screen_event.dart'                     show PageStackScreenBlocEvent;
 
-export 'src/pages/abstract.dart'                            show AbstractPage;
-export 'src/pages/material.dart'                            show AbstractMaterialPage;
-export 'src/pages/bloc_material.dart'                       show BlocMaterialPage;
-export 'src/pages/stateless_material.dart'                  show StatelessMaterialPage;
+export 'src/blocs/screen/close_event.dart'                          show ScreenBlocCloseEvent;
+export 'src/blocs/screen/configuration_changed_event.dart'          show ScreenBlocConfigurationChangedEvent;
+export 'src/blocs/screen/event.dart'                                show ScreenBlocEvent;
+export 'src/blocs/screen/screen.dart'                               show ScreenBloc;
 
-export 'src/widgets/navigator.dart'                         show PageStackBlocNavigator;
+export 'src/pages/abstract.dart'                                    show AbstractPage;
+export 'src/pages/material.dart'                                    show AbstractMaterialPage;
+export 'src/pages/bloc_material.dart'                               show BlocMaterialPage;
+export 'src/pages/stateless_material.dart'                          show StatelessMaterialPage;
+
+export 'src/widgets/navigator.dart'                                 show PageStackBlocNavigator;