+++ lib/src/blocs/page_stack/event.dart
@@ -1,2 +1,1 @@
-abstract class PageStackBlocEvent {
-}
+abstract class PageStackBlocEvent {}