+++ lib/src/blocs/screen/close_event.dart
@@ -1,4 +1,7 @@
 import 'event.dart';
 
-class ScreenBlocCloseEvent extends ScreenBlocEvent {
-}
+/// Emit this for [PageStackBloc] to remove and dispose the page.
+///
+/// The page could be anywhere in the stack as long as at least one other page
+/// is left in the stack afterwards.
+class ScreenBlocCloseEvent extends ScreenBlocEvent {}