+++ lib/src/blocs/screen/event.dart
@@ -1,2 +1,5 @@
-abstract class ScreenBlocEvent {
-}
+/// The base class for [ScreenBloc] events.
+///
+/// Extend this for your custom events like emitting notifications
+/// to be shown in overlays etc.
+abstract class ScreenBlocEvent {}