+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.1
+
+* **BREAKING**: Removed `ScreenBlocConfigurationChangedEvent.configuration`.
+* Added `AppBloc`.
+* Added `AppBlocNormalizedState`, `PageStackBlocNormalizedState`, `ScreenBlocNormalizedState`.
+* Normalize and denormalize state.
+* Create pages in `PageStackBloc`.
+
 ## 0.1.0
 
 * Initial release.